requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
Django==3.1
django-admin-sortable2==1.0.4
django-ckeditor==6.0.0
django-colorfield==0.7.0
django-easy-select2==1.5.8
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-mptt==0.13.4
Pillow==9.1.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
singlemodeladmin==0.9
six==1.16.0
sqlparse==0.4.2
Unidecode==1.3.4
